feat: refactor some code
